test: expirationd_update
Transferring the taptest to the luatest testing system with renaming the
tests and using the existing spaces in helpers.

Check if the expirationd module will work correctly on time and
after expirationd.update.
Also fixed error in task_stop_test, a bug that affected the update test.

Updated test's names:

- restart test -> test_expirationd_update

Part of #61